# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CompanyOS Core API",
    "description": "Phase 1.3 \u2014 dashboard BFF + workspace + auth + hello.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/finance/recurring": {
      "post": {
        "tags": [
          "finance-recurring"
        ],
        "summary": "POST /api/v1/finance/recurring",
        "operationId": "create_recurring",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurringRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringInvoiceDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInvoiceRequest": {
        "type": "object",
        "required": [
          "customer_id",
          "currency",
          "lines"
        ],
        "properties": {
          "base_currency": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "customer_id": {
            "type": "string",
            "description": "Finance customer public id (`cus_\u2026` projection) or sales customer id."
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineInput"
            }
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "terms": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateRecurringRequest": {
        "type": "object",
        "required": [
          "customer_id",
          "cadence",
          "next_run_at",
          "template"
        ],
        "properties": {
          "cadence": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "next_run_at": {
            "type": "string"
          },
          "template": {
            "$ref": "#/components/schemas/CreateInvoiceRequest"
          }
        }
      },
...
      "InvoiceLineInput": {
        "type": "object",
        "required": [
          "description",
          "quantity",
          "unit_price_minor"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "discount_minor": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int64"
          },
          "tax_rate_bps": {
            "type": "integer",
            "format": "int64"
          },
          "unit_price_minor": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "RecurringInvoiceDto": {
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "cadence",
          "next_run_at",
          "active"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "cadence": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "next_run_at": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}